Alternative Traces for Header Pins
Since most components will be sitting on the top side of the
device, traces for the header pins have been drawn on the bottom side
as much as possible. This should mean it will be easier to connect the
rest of the components.
A new branch has been made to preserve the former trace layout if this
version is no good.

Signed-off-by: Sophie Haynes <sophie.haynes@linaro.org>